API view is ready
This commit brings #30 along :(, so API testing needs to go on hold
because the BunqAPI is returning: Too many requests. You can do a maximum of 1 POST call per 1 second to this endpoint.

Need to create 2 form boxes to enter the userID and account ID for the enpoints that require these.

Need to adapt the API view to accept these inputs.